[REFACTOR] RestClient.Builder를 주입받아 사용하도록 수정 (#315)
* refactor:

RestClient.Builder 로 변경

* refactor:

OrderGateway 삭제 및 response 위치 수정

* refactor:

가독성 위해 코드 수정

* refactor:

PATH_SEPARATOR 상수 제거
